requirements.txt
<<<<<<< HEAD
absl-py==2.3.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
Django==5.1.5
django-cors-headers==4.3.1
django-unfold==0.62.0
djangorestframework==3.16.0
executing==2.2.0
facenet-pytorch==2.5.2
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.1
gunicorn==23.0.0
hub-sdk==0.0.20
huggingface-hub==0.20.0
hydra-core==1.3.2
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.8.0.74
packaging==24.2
pandas==2.3.0
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.16.0
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tokenizers==0.15.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
typing_extensions==4.14.0
tzdata==2025.1
ultralytics==8.1.0
urllib3==2.5.0
uuid==1.30
wcwidth==0.2.13
Werkzeug==3.1.3
=======
asgiref==3.8.1
contourpy==1.3.2
cycler==0.12.1
Django==5.1.5
django-unfold==0.62.0
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
mysqlclient==2.2.7
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
>>>>>>> 34accc8f6043820424f9169cdc62545edc287931
